+++ src/main/java/io/github/sleod/tas/common/utils/AnnotationUtils.java
@@ -1,11 +1,17 @@
 package io.github.sleod.tas.common.utils;
 
+import io.github.sleod.tas.common.enumerations.WebDriverName;
+import io.github.sleod.tas.configuration.PropertyResolver;
+import io.github.sleod.tas.core.annotations.SearchCriteria;
+import io.github.sleod.tas.exception.ExceptionBase;
+import io.github.sleod.tas.exception.ExceptionErrorKeys;
 import org.openqa.selenium.support.FindBy;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Field;
 import java.lang.reflect.Method;
 import java.lang.reflect.Proxy;
+import java.util.Arrays;
 import java.util.Map;
 
 import static io.github.sleod.tas.common.logging.SystemLogger.info;
@@ -24,24 +30,30 @@ public class AnnotationUtils {
      * @param fieldAnnotation field annotation object
      * @param config          configuration of test object with map in form {"how":"value", "using":"value"}
      * @param annotationClass annotation class static context
-     * @throws NoSuchFieldException   no such field found
-     * @throws IllegalAccessException access rights
      */
     @SuppressWarnings("unchecked")
-    public static void changeAnnotationValue(Annotation fieldAnnotation, Map<String, String> config, Class<?> annotationClass) throws NoSuchFieldException, IllegalAccessException {
-        Object handler = Proxy.getInvocationHandler(fieldAnnotation);
-        Field valueField = handler.getClass().getDeclaredField("memberValues");
-        valueField.setAccessible(true);
-        Map<String, Object> memberValues = (Map<String, Object>) valueField.get(handler);
-        //reset all value except how to default
-        Method[] methods = annotationClass.getDeclaredMethods();
-        for (Method method : methods) {
-            memberValues.put(method.getName(), method.getDefaultValue());
+    public static void changeAnnotationValue(Annotation fieldAnnotation, Map<String, String> config, Class<?> annotationClass) {
+        try {
+            Object handler = Proxy.getInvocationHandler(fieldAnnotation);
+            Field valueField = handler.getClass().getDeclaredField("memberValues");
+            valueField.setAccessible(true);
+            Map<String, Object> memberValues = (Map<String, Object>) valueField.get(handler);
+            //reset all value except how to default
+            Method[] methods = annotationClass.getDeclaredMethods();
+            for (Method method : methods) {
+                memberValues.put(method.getName(), method.getDefaultValue());
+            }
+            String how = config.get("how");
+            String using = config.get("using");
+            if (annotationClass.equals(SearchCriteria.class)) {
+                memberValues.put("locator", how + "=" + using);
+            } else {
+                memberValues.put(how, using);
+            }
+            info("modified field annotation to: " + how + " = " + using);
+        } catch (IllegalAccessException | NoSuchFieldException ex) {
+            throw new ExceptionBase(ExceptionErrorKeys.CUSTOM_MESSAGE, ex, "Exception by change annotation value in runtime: " + fieldAnnotation);
         }
-        String how = config.get("how");
-        String using = config.get("using");
-        memberValues.put(how, using);
-        info("modified field annotation to: " + how + " = " + using);
     }
 
     /**
@@ -49,13 +61,21 @@ public class AnnotationUtils {
      *
      * @param field  field with annotation
      * @param config contains how and using
-     * @throws NoSuchFieldException   for fetch private fields internal annotation handler
-     * @throws IllegalAccessException access privilege violation
      */
     @SuppressWarnings({"unchecked", "RedundantSuppression"})
-    public static void changeFindByValue(Field field, Map<String, String> config) throws NoSuchFieldException, IllegalAccessException {
-        final FindBy fieldAnnotation = field.getAnnotation(FindBy.class);
-        changeAnnotationValue(fieldAnnotation, config, FindBy.class);
+    public static void changeFindByValue(Field field, Map<String, String> config) {
+        if (PropertyResolver.getWebDriverName().equals(WebDriverName.PLAYWRIGHT.getName())) {
+            Arrays.stream(field.getAnnotations())
+                    .filter(annotation -> annotation.annotationType().equals(SearchCriteria.class)).findFirst()
+                    .ifPresent(annotation -> {
+                        changeAnnotationValue(field.getAnnotation(SearchCriteria.class), config, SearchCriteria.class);
+                    });
+        } else {
+            Arrays.stream(field.getAnnotations())
+                    .filter(annotation -> annotation.annotationType().equals(FindBy.class)).findFirst()
+                    .ifPresent(annotation -> changeAnnotationValue(field.getAnnotation(FindBy.class), config, FindBy.class));
+        }
+
     }
 
 }